Refactor main CLI code into main.py
Refactor code related to running the CLI into main.py. This refocuses
collector.py on just collector code.
